The local USNOB code inherited from the GRB pipeline had the RATCam FoV hard coded in. Has been changed to use the per instrument configs defined at the start of teh script. This only effects the on site fits since they use the local USNOB.
If you did a fir using a small FoV camera and then a large FoV with the same pointing
it was possible for the WCS to try and use the oldcached calatogue for teh smaller FoV
which would not work well. For the local USNOB this case has now been trapped and a
now USNOB file will be generated.
his only effects the on site fits since they use the local USNOB.
The same error could in principle happen on lt-archive and it is not yet caught and
automatically fixed there.